app/images.py
from dotenv import load_dotenv
from imagekitio import ImageKit
import os

load_dotenv()

# Initialize imagekit with only the private_key parameter
imagekit = ImageKit(
    private_key=os.getenv("IMAGEKIT_PRIVATE_KEY")
)