pyproject.toml
[project]
name = "cloudshare-rest-api"
version = "0.1.0"
description = "A modern, secure cloud storage and sharing platform built with FastAPI and Streamlit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.118.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "uvicorn[standard]>=0.37.0",
]
